Move header section to sections/headers folder
