+++ next/src/components/layouts/PageHeaderSections.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import SubpageListPageHeaderSection from '@/src/components/sections/SubpageListPageHeaderSection_Deprecated'
+import SubpageListPageHeaderSection from '@/src/components/sections/headers/SubpageListPageHeaderSection_Deprecated'
 import { PageHeaderSectionsFragment } from '@/src/services/graphql'
 import { isPresent } from '@/src/utils/utils'
 
@@ -10,7 +10,6 @@ type PageHeaderSectionsProps = {
 
 const PageHeaderSections = ({ sections }: PageHeaderSectionsProps) => {
   return (
-    // eslint-disable-next-line react/jsx-no-useless-fragment
     <>
       {sections?.filter(isPresent).map((section, index) => {
         // eslint-disable-next-line sonarjs/no-small-switch
